--- src/main/java/com/wanted/preonboarding/ticket/domain/entity/PerformanceSeatInfo.java
@@ -1,4 +1,0 @@
-package com.wanted.preonboarding.ticket.domain.entity;
-
-public class PerformanceSeatInfo {
-}
