+++ src/main/java/com/wanted/preonboarding/ticket/entity/Reservation.java
@@ -1,0 +1,38 @@
+package com.wanted.preonboarding.ticket.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Reservation {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    @Column(columnDefinition = "BINARY(16)", nullable = false, name = "performance_id")
+    private UUID performanceId;
+    @Column(nullable = false)
+    private String name;
+    @Column(nullable = false, name = "phone_number")
+    private String phoneNumber;
+    @Column(nullable = false)
+    private int round;
+    private int gate;
+    private char line;
+    private int seat;
+
+
+}
