--- src/main/java/com/wanted/preonboarding/ticket/infrastructure/repository/PerformanceRepository.java
@@ -1,12 +1,0 @@
-package com.wanted.preonboarding.ticket.infrastructure.repository;
-
-import com.wanted.preonboarding.ticket.domain.entity.Performance;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface PerformanceRepository extends JpaRepository<Performance, UUID> {
-    List<Performance> findByIsReserve(String isReserve);
-    Performance findByName(String name);
-}
