--- src/main/java/com/wanted/preonboarding/ticket/infrastructure/repository/ReservationRepository.java
@@ -1,12 +1,0 @@
-package com.wanted.preonboarding.ticket.infrastructure.repository;
-
-import com.wanted.preonboarding.ticket.domain.entity.Performance;
-import com.wanted.preonboarding.ticket.domain.entity.Reservation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
-    Reservation findByNameAndPhoneNumber(String name, String phoneNumber);
-}
