+++ src/main/resources/application.yml
@@ -11,7 +11,7 @@ spring:
   # JPA 설정
   jpa:
     hibernate:
-      ddl-auto: create        # DB 초기화 전략 (none, create, create-drop, update, validate)
+      ddl-auto: none        # DB 초기화 전략 (none, create, create-drop, update, validate)
     properties:
       hibernate:
         auto_quote_keyword: false
